checkpoint: renames "Attempt/factory/AttemptCreationError.ts" to "Attempt/Error/Creation.ts"
