+++ src/library/Attempt/Error/Creation.ts
@@ -1,4 +1,4 @@
-import Attempt_Error_NonActionable from '../Error/NonActionableError';
+import Attempt_Error_NonActionable from './NonActionableError';
 
 class Attempt_Error_Creation
   extends Attempt_Error_NonActionable {
