+++ src/library/Attempt/Error/assertions/assertPotentiallyActionable.ts
@@ -1,4 +1,4 @@
-import Attempt_Error_Creation from '../../factory/AttemptCreationError';
+import Attempt_Error_Creation from '../Creation';
 
 import {
   default as Attempt_Error_NonActionable,
